requirements.txt
# Core Dependencies
streamlit>=1.29.0
crewai>=0.13.0
python-dotenv>=1.0.0

# Language Models & Processing
langchain-core>=0.1.7
langchain-community>=0.0.10
langchain-huggingface>=0.0.6
groq>=0.4.2
tiktoken>=0.5.1

# Vector Database & Embeddings
langchain-chroma>=0.0.4
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Document Processing
python-magic-bin>=0.4.14
unstructured>=0.11.0
python-docx>=1.0.1
pdf2image>=1.16.3
pdfminer.six>=20221105
docx2txt>=0.8
markdown>=3.5.2
PyPDF2>=3.0.0

# Data Management
pandas>=2.2.0
openpyxl>=3.1.2

# Voice & Audio Processing
elevenlabs>=0.2.27
pydub>=0.25.1
sounddevice>=0.4.6
numpy>=1.24.0

# File & Memory Management
fsspec>=2023.12.2
aiofiles>=23.2.1
cachetools>=5.3.2

# Utilities
tqdm>=4.66.1
requests>=2.31.0
aiohttp>=3.9.1
tenacity>=8.2.3
typing-extensions>=4.9.0

# Testing & Development
pytest>=7.4.3
black>=23.12.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0